Cleaned up and refactored
