+++ src/main.js
@@ -2,7 +2,7 @@ import Vue from "vue";
 import Vuex from "vuex";
 import App from "./App.vue";
 import router from "./router";
-import store from "./store";
+import store from "./ui-store/store";
 
 Vue.config.productionTip = false;
 Vue.use(Vuex)
