--- src/store_config.js
@@ -1,3 +1,0 @@
-const free_search = "free search";
-
-export   { free_search };
