+++ src/ui-store/store_config.js
@@ -1,0 +1,3 @@
+const free_search = "Free Search";
+
+export   { free_search };
